pyproject.toml
[project]
name = "webnavix"
version = "0.1.0"
description = "Continuous Generalist Web Navigation Agent using domain-wise Mixture-of-Experts that merges individually fine-tuned LLMs as domain experts."
authors = [
  { name = "shio", email = "85730998+dino3616@users.noreply.github.com" },
]
dependencies = [
  "accelerate==0.27.2",
  "bitsandbytes==0.42.0",
  "datasets==2.19.2",
  "deepspeed==0.15.1",
  "huggingface-hub==0.24.6",
  "hydra-core==1.3.2",
  "lxml==5.3.0",
  "mergoo==0.0.10",
  "omegaconf==2.3.0",
  "peft==0.12.0",
  "python-dotenv==1.0.1",
  "torch==2.4.1",
  "tqdm==4.66.2",
  "transformers==4.42.4",
  "trl==0.10.1",
  "wandb==0.17.9",
  "weblinx[eval]==0.3.0",
]
readme = "README.md"
requires-python = "~=3.12"

[tool.rye]
managed = true
dev-dependencies = ["lefthook==0.1.2", "ruff==0.6.4"]

[tool.rye.scripts]
check = { chain = ["lint", "fmt"] }
lint = "ruff check ./ --diff"
fmt = "ruff fmt ./"
